Fix. build  시 낫던 에러 수정

+++ pmy/src/app/utils/service/localStorageService.ts
@@ -1,6 +1,7 @@
 export class LocalStorageService {
-    // 데이터 저장
-    static setItem<T>(key: string, value: T): void {
+  // 데이터 저장
+  static setItem<T>(key: string, value: T): void {
+    if (typeof window !== 'undefined') {
       try {
         const serializedValue = JSON.stringify(value);
         localStorage.setItem(key, serializedValue);
@@ -8,9 +9,11 @@ export class LocalStorageService {
         console.error(`Error saving data to localStorage with key "${key}":`, error);
       }
     }
-  
-    // 데이터 가져오기
-    static getItem<T>(key: string): T | null {
+  }
+
+  // 데이터 가져오기
+  static getItem<T>(key: string): T | null {
+    if (typeof window !== 'undefined') {
       try {
         const serializedValue = localStorage.getItem(key);
         if (serializedValue === null) {
@@ -22,28 +25,36 @@ export class LocalStorageService {
         return null;
       }
     }
-  
-    // 데이터 삭제
-    static removeItem(key: string): void {
+    return null;
+  }
+
+  // 데이터 삭제
+  static removeItem(key: string): void {
+    if (typeof window !== 'undefined') {
       try {
         localStorage.removeItem(key);
       } catch (error) {
         console.error(`Error removing data from localStorage with key "${key}":`, error);
       }
     }
-  
-    // 모든 데이터 초기화
-    static clear(): void {
+  }
+
+  // 모든 데이터 초기화
+  static clear(): void {
+    if (typeof window !== 'undefined') {
       try {
         localStorage.clear();
       } catch (error) {
         console.error("Error clearing localStorage:", error);
       }
     }
-  
-    // 특정 키가 존재하는지 확인
-    static hasKey(key: string): boolean {
+  }
+
+  // 특정 키가 존재하는지 확인
+  static hasKey(key: string): boolean {
+    if (typeof window !== 'undefined') {
       return localStorage.getItem(key) !== null;
     }
+    return false;
   }
-  
+}
